requirements.txt
asgiref==3.8.1
bootstrap4==0.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
distlib==0.3.8
Django==5.0.6
django-cors-headers==4.9.0
django-crispy-forms==1.14.0
djangorestframework==3.15.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
filelock==3.15.4
idna==3.7
mercadopago==2.2.1
numpy==2.2.5
openpyxl==3.1.5
pandas==2.2.3
pillow==10.3.0
platformdirs==4.2.2
PyJWT==2.10.1
pymssql==2.3.4
pypyodbc==1.3.6
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.31.0
setuptools==69.5.1
six==1.17.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.26.3
